Changed data type for parameter avg.

+++ calculation.hpp
@@ -1,6 +1,6 @@
 #include <iostream>
 using namespace std;
-void calculation(int a, int b, int &sum, int &avg){
+void calculation(double a, double b, int &sum, double &avg){
     sum = a + b;
     avg = (a + b)/2;
 }
